Improve In-app updater
Adjust the feature so that it works for Android 11 which has
introduces permission access. Also updates will be checked 6 hours
in order to reduce bandwidth consumption. File will be deleted
once update is successful.

+++ app/src/main/java/com/whatsapp/Broadcast.java
@@ -3,9 +3,11 @@ package com.whatsapp;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
+import android.widget.Toast;
 
 import androidx.core.content.FileProvider;
 
@@ -17,7 +19,6 @@ import x.Update;
 
 public class Broadcast extends BroadcastReceiver {
 
-    File file = new File(Environment.getExternalStorageDirectory() + Constant.folder + Constant.file);
     public void onReceive(Context ctxt, Intent intent) {
         if (did() == intent.getLongExtra("extra_download_id", -1)) {
             installfile(ctxt);
@@ -29,20 +30,33 @@ public class Broadcast extends BroadcastReceiver {
     public void installfile(Context context) {
         Intent intent = new Intent();
         intent.setAction("android.intent.action.VIEW");
+        File file = new File(context.getExternalFilesDir(null) + Constant.dots + Environment.DIRECTORY_DOWNLOADS, Constant.file);
         if (Build.VERSION.SDK_INT >= 24) {
-            Uri uri = FileProvider.getUriForFile(context.getApplicationContext(), context.getApplicationContext().getPackageName() + ".provider", this.file);
+            Uri uri = FileProvider.getUriForFile(context.getApplicationContext(), context.getApplicationContext().getPackageName() + ".provider", file);
             intent.setDataAndType(uri, Constant.type);
             intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION | Intent.FLAG_ACTIVITY_NEW_TASK);
         } else {
-            intent.setDataAndType(Uri.fromFile(this.file), Constant.type);
+            intent.setDataAndType(Uri.fromFile(file), Constant.type);
         }
         context.startActivity(intent);
     }
 
-    public static void delete() {
-        File file = new File(Environment.getExternalStorageDirectory() + Constant.folder + Constant.file);
-        if (file.exists())
-        file.delete();
+    public static void updated(Context context) {
+        File file = new File(context.getExternalFilesDir(null) + Constant.dots + Environment.DIRECTORY_DOWNLOADS, Constant.file);
+        int vercode = Main.sharedPreferences.getInt("vercode",0);
+        int vercodeofapp = 0; boolean b = false;
+        try {
+            vercodeofapp = context.getPackageManager().getPackageInfo(context.getPackageName(), 0).versionCode;
+        } catch (PackageManager.NameNotFoundException ignored) {
+        }
+
+        if(vercodeofapp > vercode){
+            if (file.exists())
+                b = file.delete();
+            Main.sharedPreferences.edit().putInt("vercode", vercodeofapp).apply();
+            if (b)
+                Toast.makeText(context, Constant.updsuccess, Toast.LENGTH_SHORT).show();
+        }
     }
 
     private long did() {
